requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0.0
argon2-cffi>=23.1.0
PyJWT>=2.8.0
